+++ spec/unit/rails/railtie_spec.rb
@@ -4,12 +4,16 @@ RSpec.describe "Chronos Rails Railtie" do
   it "registers installation after application configuration without Zeitwerk" do
     railtie_base = Class.new do
       class << self
-        attr_reader :initializer_name, :initializer_options, :initializer_block
+        attr_reader :initializer_name, :initializer_options, :initializer_block, :rake_tasks_block
 
         def initializer(name, options = {}, &block)
           @initializer_name = name
           @initializer_options = options
           @initializer_block = block
+        end
+
+        def rake_tasks(&block)
+          @rake_tasks_block = block
         end
       end
     end
@@ -23,6 +27,7 @@ RSpec.describe "Chronos Rails Railtie" do
     expect(Chronos::Rails::Railtie.superclass).to equal(railtie_base)
     expect(Chronos::Rails::Railtie.initializer_name).to eq("chronos.install")
     expect(Chronos::Rails::Railtie.initializer_options).to eq(:after => :load_config_initializers)
+    expect(Chronos::Rails::Railtie.rake_tasks_block).to be_a(Proc)
   end
 
   it "is discovered automatically when Rails is already loaded" do
@@ -30,6 +35,7 @@ RSpec.describe "Chronos Rails Railtie" do
       module Rails
         class Railtie
           def self.initializer(*_arguments); end
+          def self.rake_tasks; end
         end
       end
       require "chronos"